+++ services/server/test/chains/chain-tests.spec.ts
@@ -1742,6 +1742,14 @@ describe("Test Supported Chains", function () {
     "shared/",
   );
 
+  // Quantum Portal Network
+  verifyContract(
+    "0x511f9B726559384f10371Bc7f19A925F864a9997",
+    "26100",
+    "Quantum Portal Network",
+    "shared/",
+  );
+
   it("should have included Etherscan contracts for all testedChains having etherscanAPI", function (done) {
     const missingEtherscanTests: ChainApiResponse[] = [];
     supportedChains